Fix to allow ppc64le build on x86_64

+++ symtabAPI/src/emitElfStatic-ppc64.C
@@ -337,7 +337,7 @@ bool emitElfStatic::archSpecificRelocation(Symtab* targetSymtab, Symtab* srcSymt
 	  assert(0);
 	}
         rewrite_printf("\tbefore: relocation = 0x%lx @ 0x%lx target data %lx %lx %lx %lx %lx %lx \n", 
-	(unsigned long)relocation, relOffset,targetData[dest-2],  targetData[dest-1], targetData[dest], targetData[dest+1],  targetData[dest+2],  targetData[dest+3]);
+	(unsigned long)relocation, relOffset,(unsigned long)targetData[dest-2],  (unsigned long)targetData[dest-1], (unsigned long)targetData[dest], (unsigned long)targetData[dest+1],  (unsigned long)targetData[dest+2],  (unsigned long)targetData[dest+3]);
 
 	if (relocation_length == 64) {
 		char *td = (targetData + dest - (dest%8));
@@ -350,7 +350,7 @@ bool emitElfStatic::archSpecificRelocation(Symtab* targetSymtab, Symtab* srcSymt
         	memcpy(td, &target, sizeof(Elf64_Word));
 	}
 
-        rewrite_printf("\tafter: relocation = 0x%lx @ 0x%lx target data %lx %lx %lx %lx %lx %lx \n", (unsigned long)relocation, relOffset,targetData[dest-2],  targetData[dest-1], targetData[dest], targetData[dest+1],  targetData[dest+2],  targetData[dest+3]);
+        rewrite_printf("\tafter: relocation = 0x%lx @ 0x%lx target data %lx %lx %lx %lx %lx %lx \n", (unsigned long)relocation, relOffset,(unsigned long)targetData[dest-2],  (unsigned long)targetData[dest-1], (unsigned long)targetData[dest], (unsigned long)targetData[dest+1],  (unsigned long)targetData[dest+2],  (unsigned long)targetData[dest+3]);
     } else if (PPC32_WIDTH == addressWidth_ ){
 	int relocation_length = sizeof(Elf32_Word)*8; // in bits
 	int relocation_pos = 0; // in bits
@@ -645,7 +645,7 @@ default:
      return false;
         }
 
-        rewrite_printf(" relocation = 0x%lx @ 0x%lx target data 0x%lx %lx %lx %lx \n", relocation, relOffset, targetData[dest], targetData[dest+1],  targetData[dest+2],  targetData[dest+3]);
+        rewrite_printf(" relocation = 0x%lx @ 0x%lx target data 0x%lx %lx %lx %lx \n", relocation, relOffset, (unsigned long)targetData[dest], (unsigned long)targetData[dest+1],  (unsigned long)targetData[dest+2],  (unsigned long)targetData[dest+3]);
 	if (rel.getRelType() == R_PPC_REL24) {
 	unsigned int *td = (unsigned int *) targetData;
 	unsigned int target;
@@ -658,7 +658,7 @@ default:
 	target = td[dest/4];
 	target = setBits(target, relocation_pos, relocation_length, relocation);
         memcpy(&td[dest/4], &target, sizeof(Elf32_Word));
-        rewrite_printf(" relocation = 0x%lx @ 0x%lx target data 0x%lx %lx %lx %lx \n", relocation, relOffset, targetData[dest], targetData[dest+1],  targetData[dest+2],  targetData[dest+3]);
+        rewrite_printf(" relocation = 0x%lx @ 0x%lx target data 0x%lx %lx %lx %lx \n", relocation, relOffset, (unsigned long)targetData[dest], (unsigned long)targetData[dest+1],  (unsigned long)targetData[dest+2],  (unsigned long)targetData[dest+3]);
 	}
     if (branch_pred >= 0) {
 	unsigned int *td = (unsigned int *) targetData;
